add tripleo scenario004 to experimental queue
A recent change [1] in master added a tripleo scenario004
job to the experimental queue so we can run it before merging
changes that might affect downstream tripleo CI.  We can't
cherry-pick that change because that scenario isn't supported
on single node CI in stable/rocky.

Here we add a multinode version of that scenario.

Change-Id: I2743ce3c472d940c23baf3955e09a81427e381e1