Include .inc files in doc8 linting
This issue was recently seen in cinder when the gate failed
to detect trailing whitespaces in '.inc' files.
As seen, manila was also prone to such issues but thankfully
no error was found in '.inc' after running against doc8.
It still seems important to include it in the testing to avoid
future similar issues.

Change-Id: Ie1fd4ddb4448c05559a31098302f84367fe6ed52