[pylint] Fix lintstack.py for python2.x
tools/lintstack.py helper is used to run
pylint on manila in all branches prior to
the Stein release. In some downstream testing
with python2.x, it was found to be using a
version of the print() function available only
in python3x.

We can enable the python3 style print function
from __future__. See [1] for how this import
is dealt with.

[1] https://docs.python.org/2/reference/simple_stmts.html#future

Depends-on: https://review.openstack.org/642715
Change-Id: I41c93ef6c0472ae4293a733bb0ea99511a686440