Bump timeout on sqlalchemy migration test
As we add more migrations in the database, the
time taken by the test increases. The test
`manila.tests.db.migrations.alembic.
test_migration.TestManilaMigrationsPostgreSQL.test_walk_versions`
runs upgrades, downgrades and database checks for
every migration. It has been timing out on numerous
occasions because it can't run under 5 minutes.

We feel the pinch of the timeouts because
the test itself is run as part of multiple
jobs (ex: unit test jobs on py2x, py3x, coverage,
requirements/lower-constraints check) and
can fail anywhere because it was running
on a "slow" node.

Change-Id: I1fdef6e907110681cb5ffb59f6f90f8eb67f118b
Partial-Bug: #1793364