Port dummy driver manage/unmanage changes to stable
Branchless manila-tempest-plugin had changes for
the manage share server feature that required changes in
the dummy driver for its manage and unmanage routines.

Here we test porting the dummy driver changes for the
manage and unmanage share and snapshot methods, omitting
the manage and unmanage methods for share servers since
these are not supported in the stable branches.

Note that backporting such changes to a stable branch
for a regular driver would likely be unacceptable given
stable policy but the dummy driver is test-only code that
never runs outside of tests.

Change-Id: Ib9e5dd3ad6b745ea1dfe442d9357a70ca4c148ec
(cherry picked from commit 6b25b0a4075550163554895b5a1e275675ef9f21)