Fix server delete attempt along with share net deletion
This bug was introduced when manage/unmange servers with DHSS=True
was merged. This behavior is expected except in two scenarios:

1. If the share server was managed;
2. If any share belonging to this share server was unmanaged;

In both cases, the `is_auto_deletable` field is set to False. This
field was added to prevent the share server from being deleted by
the periodic cleanup job.

This patch fixes the bug by checking the `is_auto_deletable` field
when the share network deletion triggers the share server deletion.

APIImpact

Change-Id: Ib83b8222100fd5a3d718ad2cb7887b1d79ebd546
Depends-On: I7d330b8b5fb3f08d3cbdae2c1735e266e75a70d3
Closes-bug: #1820118