Dummy driver: Don't fail unmanage on malformed share servers
When a share server is managed with invalid details, since
we cannot find backend network allocations in the dummy driver,
we raise an error on manage, setting the status to "manage_error".
Such a share server can be "unmanaged" or deleted, and we shouldn't
fail this request, either case.

Change-Id: I4144a532d5827b929a1dce0d4d642e6d1cee47b1
Related-Bug: #1820118