INFINIDAT: suppress 'no-member' pylint errors
The INFINIDAT 'infinisdk' module uses the 'capacity' module to describe
capacity sizes (as similar to the 'oslo_units' module). As the module
might not be present in its testing environment, pylint has a hard time
recognizing some of the module's members. This change suppresses the
pylint errors only where needed, as the 'capacity' module is used only
in very specific methods in the driver.

Change-Id: Ife41e05c743a8f21ff2a91c0c3b9999b1cc25245
Closes-Bug: #1816646