Conditionally restore default route in setup_ipv6
Restoring the default IPv6 route is not needed in CI and
fails when there are multiple defaults, but it is useful
in local devstacks where multiple default routes are not
typical.

Add a variable in settings and use it to make this behavior
conditional and set it to False for the lvm job.

Closes-bug: #1836788
Change-Id: Id73de8100509ec5935641f5f35f93f482d108bcd
(cherry picked from commit 70e908344dd4db81652264a2de2c04ce9666bb46)
(cherry picked from commit 99c0a110920910bfd379d544e7085daca90b1132)