Skip NFS/Samba install for CephFS
Today we either expose the CephFS back end via the CephFS native
protocol, in which case kernel NFS and Samba are not needed, or
by NFS-Ganesha, in which case Samba is not needed and kernel NFS
is incompatible (we just turn it off again later).

In the future someone could develop support for exposing CephFS via
Samba so we'd add specific checks for that then and install Samba.

Change-Id: Id65ce2536a3deb32a04a70bd0c49081297741965