Add asynchronous error info into messages when share extend error
The database table messages is used to record various asynchronous error
messages, extend share is an asynchronously invoked procedure, when the
driver layer fails to execute extend share, we should record this
asynchronous error into database. This commit modify this behavior, and
now this asynchronows error is recorded in the database.

Closes-Bug:#1850264
Change-Id: I06a9bf24ab3d4aa5f4f5446d344bf55d71803e47