Use psycopg2-binary for test-requirements
In order to run PostgreSQL tests, we need to have the PostgreSQL
dependency available on the system.  We currently use the dependency
which forces building the Python package from source.  That makes
running the tests locally non-trivial and requiring the installation
of packages locally.

This patch moves that dependency to the binary one so that we can
simply deploy a binary artifact during CI, which simplifies the
abilitiy to run those tests.

Change-Id: I33369712370b86feca843f118f35063d0993adcb
(cherry picked from commit 3e82f118b7f7d8094a7e65260a3280a13c4b4df1)