[stable-only] Pin neutron-tempest-plugin to 0.9.0
manila's IPv6 jobs setup bgp with the help of
neutron-dynamic-routing. The jobs enable
neutron-dynamic-routing using its devstack
plugin to install it alongside openstack/neutron.

neutron-tempest-plugin is a "branchless" tempest
plugin to the neutron project. We don't need to
install or run this tempest plugin in manila test
jobs. However, it is imported by test code within
the neutron-dynamic-routing project, forcing us to
include it in the list of requirements for the IPv6
jobs.

neutron-tempest-plugin dropped support for python3
in ussuri. python3 only code was added to
neutron-tempest-plugin causing failures in the
voting ipv6 lvm test job
(manila-tempest-minimal-dsvm-lvm).

So, pin the version of neutron-tempest-plugin we're
pulling in to a tagged git ref when the plugin last
supported python2.

Change-Id: I82854d0f9075e6fc7d33c6459d71d7f0d502ff3c
Closes-Bug: #1868905
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>