Prevent share type deletion if linked to group types
Share types can be associated with share group
types. This relationship needs to be dropped in
case share types are to be deleted, or we'll be
leaving behind dead references in the database.
Share groups cannot exist without share group
types, and hence we don't need to check on existing
groups when deleting share types; checking for
existing group types will suffice.

Change-Id: Id86a93f1923a493cefe73e2d90d5094005b42ae4
Closes-Bug: #1699836
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>