[devstack][ci] Stop requiring g-reg service
When using manila's devstack plugin, we pull down
the latest manila testing image created with the
manila-image-elements project, and upload it to the
stack's Glance/Image service.

Glance's V2 API no longer needs g-reg which was a
legacy database proxy service for the older glance
API. Our devstack plugin should not have a hard
dependency on this service.

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-May/094773.html
[2] https://specs.openstack.org/openstack/glance-specs/specs/newton/approved/glance/deprecate-registry.html

Change-Id: I58603032fb215ac0be2f76c8cd9fbce775be41dd
Closes-Bug: #1859847