Enable cephfs snapshots
The CephFS driver optionally advertised snapshot
support to manila, based on a configuration
option "cephfs_enable_snapshots". Manila does not
provide snapshots by default to users, unless an
administrator chooses to set the "snapshot_support"
extra-spec in the share type they're using. So this
configuration option is really an unnecessary safeguard
at this point.

Moreover, CephFS snapshots are fully supported from
Ceph Nautilus, and with Id847b1b5e6623172f2600afd69e51182067444cb,
manila added a support statement for the Victoria
release that indicates that Nautilus is the minimum
version of ceph supported. So let's default this
option to True, and deprecate this option at the same
time to eliminate this duplicate safeguard.

Partially-implements: bp update-cephfs-drivers
Depends-On: I885ff78586540d01590f98cca2a00d55161dbe04
Change-Id: I1ef42b414f8af0b8f9a912655757ec200a75c309
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>