Update NFS helper restart
Because ``service`` commands are deprecated in favor of
``systemctl` commands, the check to see if the service
command succeeds in the NFS helper used by e.g. the
lvm and generic back ends fails.  The string sought to
indicate an error is never present.

Update the helper to use ``systemctl`` commands instead
and error check for them correctly.

Closes-bug: #1887695
Change-Id: Ibb1c15f997fa60e1ba5bfb7437502d8aa454d35f