[ci] Re-enable scenario tests for lvm job
We temporarily disabled scenario tests in
the voting LVM job [1] because of a bug
in which a kernel problem caused a reboot
in the middle of the job when it ran in a
Xen virtualization environment on rax nodes.

The kernel has been updated on those nodes
and the problem has gone away, as evidenced
in various checks jobs in this review, so
let's reenable the scenario tests.
    .
[1] https://review.opendev.org/#/c/740507/
[2] https://launchpad.net/bugs/1886988

Closes-bug: #1886988

Change-Id: I501d0b6537653613a58ea1fe606f6ef66b8b6d38