fix unmange share with manage_error status will lead to quota error
if we failed to manage a share, we don't need to commit the quota usages. so
we should skip quota usages cuts when delete or unmange the share with
status of "error_manage". and the size of error_manage share should be
zero.
we should depends on manila-tempest with max_api_microversion is 2.57,
below depends id is the is the latest version number before the 2.61
upgrade.

Closes-Bug:#1883506

Change-Id: I5c81dd6780890c55c8c6a92491c3f4f507531fdb
(cherry picked from commit 73d0a74f3b7c93e0be8779c959b74bdb5b9fd57f)