[extended-maintenance-branches-only] Replace LVM job
the voting LVM job (manila-tempest-minimal-dsvm-lvm-centos-7)
has been failing for a while due to the issue identified
in [1]. The fix for the issues didn't suffice for CentOS,
and we use CentOS because of missing IPv6 related fixes
in Xenial's nfs-utils package. to avoid impacting helpful
backports to these branches, this commit introduces a
new voting job called "manila-tempest-dsvm-lvm-em-only"
that deploys manila with the LVM driver on Xenial and
runs IPv4 scenario tests.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015432.html
Related-Bug: #1888717
Change-Id: I4a9ae0744757d1c9bfe4f2b213eb04364992af0d
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>