[ci] Remove explicit compression of log files
Compression is not needed post [1], explicit
compression renders them non-browsable.

This commit applies to legacy style jobs that
will soon be replaced with zuulv3 native jobs;
however, fixing it isn't terrible, and we
can use this fix for stable branches
where zuulv3 native jobs may not come for a
while.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-January/011875.html
Co-Authored-By: yatinkarel <ykarel@redhat.com>

Change-Id: Icd4f9dc435a807f6f729e4eb2d4df4b282ec79ac
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>