Handle oslo.messaging ping endpoint
We have a test that attempts to emulate a messaging endpoint and
fails if anything other than the endpoint's target property is
inspected. With the addition of the ping endpoint in
I51cf67e060f240e6eb82260e70a057fe599f9063, this now blows up
with the latest oslo.messaging release.

To handle this situation, our fake class just needs to also
check for the oslo_rpc_server_ping attribute.

Change-Id: I485c1724614abf03928b83280c88ffc4cdfdb1c4
Co-Authored-By: Sean McGinnis <sean.mcginnis@gmail.com>
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>