[NetApp] Fix access rules for CIFS promoted replica
This patch fixes the access rules for NetApp promote replica when
using CIFS protocol. When promoting a replica, the NetApp ONTAP
driver updates the access rules for the promoted CIFS share entity
before actually creating it, failing on having those rules
applied.

The bug is fixed by switching the order of updating the access
and creating the promoted CIFS share entity.

Change-Id: I60e4057dc962d96cff57dea88587a28c2043b499
Closes-Bug: #1896949