Implement secure RBAC for share group types
This commit updates the policies for
share group types to understand scope
checking and account for a read-only role.
This is part of a broader series of
changes across OpenStack to provide a
consistent RBAC experience and improve
security.

Also fix the HTTP method in the policy
doc for this API.

Change-Id: I8b95e1e70f74052e5bd4af1ba29842420bafd0b2
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>