[stable/train] Update requirements and constraints
pip 20.3 brings in a strict dependency resolver which
is enabled by default. This dependency resolver
unearthed the error in our tox file where we
were constraining our lower-constraints job with both
upper and lower constraints files.

The upper constraint file was being pulled in because
we had an unnecessary "pip install" command added as
"install_command". We have removed the use of
install_command from ussuri [1]; but the older branches
need to be addressed.

As a result of fixing our tests, we're also
having to update the lower bounds of some
requirements on this branch.

Version changes for direct dependencies:
* eventlet (0.18.2 --> 0.24.1)
(0.25.2 is the upper-constraint)
* oslo.messaging (5.29.0 --> 6.4.1)
(6.4.1 was released 2018 June, during Rocky [4]
upper-constraint is 10.2.2)
oslo.service (1.24.0 --> 1.31.0)
(1.31.0 was released 2018 April, during Rocky [5]
upper-constraint is 1.40.2)
* psycopg2-binary (2.6.2-->2.8.3)
(psycopg2-binary is a unit test requirement)
* bashate (0.5.0 --> 0.6.0)
(bashate is a linting test requirement)

Version changes for transitive dependencies:
* MarkupSafe (1.0 --> 1.1.1)
(1.1.1 is the upper-constraint)
* psutil (5.4.3 --> 5.6.3)
(5.6.3 is the upper-constraint)

[1] I6344f44fb955d631a4f0a13fa53ecc19826c7ea6
[2] http://pyfound.blogspot.com/2020/11/pip-20-3-new-resolver.html
[3] https://pip.pypa.io/en/stable/user_guide/#changes-to-the-pip-dependency-resolver-in-20-3-2020
[4] http://lists.openstack.org/pipermail/release-announce/2018-June/005341.html
[5] http://lists.openstack.org/pipermail/release-announce/2018-April/004759.html

Change-Id: I5e6b855e8c8d7ffed15b6e419169067e1f04a9cb
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>