[stable/rocky] Adjust CI jobs
stable/rocky was first released in August 2018,
and the final release from this branch (7.4.1)
was made on 12 Mar 2020 [1]. This branch is
currently open only to take necessary third
party driver fixes and maintenance is at best
effort. A number of CI jobs have been broken
on this branch for a while. Let's clean them
up so we don't waste CI resources when backports
are proposed. This definitely implies loss
of test coverage, however, we have to weigh
the cost of repairing these CI jobs against
momentum of maintenance and work for the
future branches.

Jobs being dropped are:

* openstack-lower-constraints
pip's latest dependency resolver unearthed a bunch
of broken dependencies on this branch. Resolving these
would involve bumping the "lower" requirements, the
team does not think this will be useful for anyone,
since we will not be altering requirements for this
branch anymore.
* manila-tempest-dsvm-mysql-generic and
The reference generic driver had four jobs on this
branch, manila-tempest-dsvm-generic-no-share-servers.
We can cut it down to two by dropping coverage
for the DHSS=False configuration and multibackend,
both of which are tested in the other jobs, albeit
with other backends.
* manila-tempest-minimal-dsvm-lvm-centos-7
This is a lost cause, we don't expect to fix
devstack on centos-7 anytime soon for these
old extended maintenance branches (See bug#1888717)
* manila-rally-multibackend and
manila-rally-multibackend-no-ss - these rally
jobs have been broken for a while, there's
barely any reason to think we'll cause performance
regressions in the API on this branch and use
these jobs to judge them.

Also drop all the experimental jobs and remove
"openstack-tox-cover" from the post queue.

[1] http://lists.openstack.org/pipermail/release-announce/2020-March/008644.html

Change-Id: Ice936853ae6b09e4ce4d5df6e6b27670cbb8f0d6
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>