Fix CI issues and adjust test jobs
It is incorrect to install tempest and manila-tempest-plugin
globally in devstack jobs. So this commit removes the use
of the test plugin's devstack installation hook, and uses
a "tempest-venv" that devstack sets up to run the tests.

To stay consistent with later branches, we enhance
the devstack plugin.sh script to set api versions
as tempest configuration options.

Drop Rally testing since the jobs are
broken and no API perf regressions
are expected to be caught on this branch
that we don't also see in newer maintained
branches.

We don't lose much test coverage if we drop
some test jobs:

- manila-tempest-dsvm-postgres-generic-singlebackend
"single backend" jobs are not very useful.
postgresql is tested elsewhere

- manila-tempest-dsvm-generic-no-share-servers
We have other DHSS=False reference drivers, don't
need to test this generic driver job on every commit.
We can move this to the experimental queue so we
can call on it if necessary.

- manila-tempest-dsvm-glusterfs-nfs-heketi
 and manila-tempest-dsvm-glusterfs-native-heketi
Neither of these jobs are maintained, and they've
been unused in the experimental queue.

So drop these jobs, and delete some playbooks
and code for the unused/orphaned test jobs.

Depends-On: I2eceb1d5fe42b1b779e115a0144a5d9639d3753f
Change-Id: I833812e04a4649283bdc619ad5278217aa0e3c23
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>